refactor(project): complete CLI modularization and documentation overhaul
This major refactoring improves the project's long-term maintainability,
testability, and user experience by restructuring the CLI codebase and
rewriting the documentation.

Key Changes:
- Code Refactoring:
  - 将单体 `main.py` 模块化为清晰的 `src-layout`，并为每个命令组建立独立模块
    (例如：`cli/user.py`、`cli/space.py`)。
  - 将所有面向开发者的命令 (`test`、`coverage`、`pkg`、`artifacts`) 统一整合
    到新的 `dev` 子命令下。
  - 抽取所有通用的 helper 函数到 `cli/utils.py`。

- Documentation Overhaul:
  - 重写 `README.md`，作为面向用户的“5分钟快速上手”指南。
  - 将所有面向开发者的信息 (构建、测试、发布流程) 迁移到新的、完整的
    `CONTRIBUTING.md`。
  - 新增 `CODE_OF_CONDUCT.md` 和 `SECURITY.md`。
  - 将所有设计/日志文档迁移并更新到集中化的 `docs/` 目录。

- Testing & Robustness:
  - 修复了 `demo quickstart` 命令中的关键 bug，使其在缺失 C 二进制文件时也
    能正常运行。
  - 将 `integration_space.sh` 脚本集成到默认的 `make coverage` 测试套件中，
    填补了主要的测试空缺。
  - 显著提升了核心模块 (如 `client`、`demo`、`space`) 的 Python
    单元测试覆盖率。
