+++ docs/CODE_OF_CONDUCT.md
@@ -1,0 +1,16 @@
+# 行为准则（Code of Conduct）
+
+本项目采用简化版的 Contributor Covenant 行为准则。
+
+## 我们的标准（Our Standards）
+- 尊重且包容。
+- 禁止骚扰、人身攻击与仇恨言论。
+- 使用友好、包容的语言。
+- 假定善意，寻求建设性解决方案。
+
+## 执行（Enforcement）
+- 请通过仓库维护者或 GitHub 举报渠道私下报告事件。
+- 维护者可采取适当措施，包括锁定讨论、临时或永久限制参与等。
+
+## 归属（Attribution）
+基于 Contributor Covenant 2.x（简化版）。
