+++ docs/SECURITY.md
@@ -1,0 +1,9 @@
+# 安全策略（Security Policy）
+
+## 报告安全问题（Reporting a Vulnerability）
+- 请通过仓库维护者或 GitHub Security Advisories 私下报告安全问题。
+- 提供可复现步骤、受影响版本与潜在影响评估，便于快速定位与修复。
+
+## 披露（Disclosure）
+- 目标在 7 天内确认收到并在 30 天内给出修复计划或状态更新。
+- 在修复发布前，请勿公开披露细节。
